[DEV-1298] Formik refactor
