+++ src/pages/get-funded/index.us.js
@@ -1,7 +1,6 @@
 import React from "react"
 import styled from "styled-components"
-import { Formik } from "formik"
-import { graphql, navigate } from "gatsby"
+import { graphql } from "gatsby"
 // import { GatsbyImage, getImage } from "gatsby-plugin-image"
 
 import LangLayout from "./../../layouts/us"
@@ -10,13 +9,10 @@ import LogoWhite from "./../../images/svg/Outfund-logo-white.svg"
 import LogoDark from "./../../images/svg/Outfund-logo.svg"
 import LogoStrip from "./../../images/svg/get-funded-logo-strip.svg"
 
-import { Button, Link } from "./../../components"
-import F from "./../../styles/new/form"
+import { Link, GetFundedForm } from "./../../components"
 import T from "../../styles/new/typography"
 import { theme } from "./../../styles/new/theme"
 import { Triangle } from "./../../styles/utils"
-
-import { VisuallyHidden } from "./../../styles/utils"
 
 export const query = graphql`
   query usGetFundedPage {
@@ -40,12 +36,6 @@ export const query = graphql`
     }
   }
 `
-
-const encode = (data) => {
-  return Object.keys(data)
-    .map((key) => encodeURIComponent(key) + "=" + encodeURIComponent(data[key]))
-    .join("&")
-}
 
 const GetFunded = (props) => {
   const data = props.data.getFundedJson.content
@@ -85,167 +75,7 @@ const GetFunded = (props) => {
                 <Title>{data.title}</Title>
                 <Description>{data.description}</Description>
               </TextWrapper>
-
-              <FormWrapper>
-                <Formik
-                  initialValues={{
-                    name: "",
-                    email: "",
-                    phone: "",
-                    website: "",
-                    amr: "",
-                  }}
-                  onSubmit={(values, actions) => {
-                    fetch("/", {
-                      method: "POST",
-                      headers: {
-                        "Content-Type": "application/x-www-form-urlencoded",
-                      },
-                      body: encode({ "form-name": "get-funded-us", ...values }),
-                    })
-                      .then(() => {
-                        navigate("/us/thank-you/")
-                        actions.resetForm()
-                      })
-                      .catch(() => {
-                        alert("Error")
-                      })
-                      .finally(() => actions.setSubmitting(false))
-                  }}
-                  validate={(values) => {
-                    const emailRegex =
-                      /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i
-                    const errors = {}
-                    if (!values.name) {
-                      errors.name = "Name Required"
-                    }
-                    if (!values.email || !emailRegex.test(values.email)) {
-                      errors.email = "Valid Email Required"
-                    }
-                    if (!values.phone) {
-                      errors.phone = "Phone Number Required"
-                    }
-                    if (!values.website) {
-                      errors.website = "Company Website Required"
-                    }
-                    if (!values.amr) {
-                      errors.amr = "Monthly Revenue Required"
-                    }
-                    return errors
-                  }}
-                >
-                  {({
-                    values,
-                    errors,
-                    touched,
-                    handleChange,
-                    handleBlur,
-                    handleSubmit,
-                    isSubmitting,
-                    isValid,
-                    dirty,
-                  }) => (
-                    <F.FormikForm
-                      name="get-funded-us"
-                      data-netlify={true}
-                      netlify-honeypot="bot-field"
-                    >
-                      <VisuallyHidden>
-                        <label>
-                          Don’t fill this out if you’re human:
-                          <input name="bot-field" tabIndex="-1" />
-                        </label>
-                      </VisuallyHidden>
-                      <F.Group>
-                        <VisuallyHidden>
-                          <label htmlFor="name">{data.form.name}</label>
-                        </VisuallyHidden>
-                        <F.FormikField
-                          $valid={errors.name && touched.name}
-                          name="name"
-                          placeholder={data.form.name}
-                        />
-                        <F.ErrorWrapper>
-                          <F.FormikError component="div" name="name" />
-                        </F.ErrorWrapper>
-                      </F.Group>
-                      <F.Group>
-                        <VisuallyHidden>
-                          <label htmlFor="email">{data.form.email}</label>
-                        </VisuallyHidden>
-                        <F.FormikField
-                          $valid={errors.email && touched.email}
-                          name="email"
-                          placeholder={data.form.email}
-                        />
-                        <F.ErrorWrapper>
-                          <F.FormikError component="div" name="email" />
-                        </F.ErrorWrapper>
-                      </F.Group>
-                      <F.Group>
-                        <VisuallyHidden>
-                          <label htmlFor="message">{data.form.phone}</label>
-                        </VisuallyHidden>
-                        <F.FormikField
-                          $valid={errors.phone && touched.phone}
-                          name="phone"
-                          placeholder={data.form.phone}
-                        />
-                        <F.ErrorWrapper>
-                          <F.FormikError component="div" name="phone" />
-                        </F.ErrorWrapper>
-                      </F.Group>
-                      <F.Group>
-                        <VisuallyHidden>
-                          <label htmlFor="message">{data.form.website}</label>
-                        </VisuallyHidden>
-                        <F.FormikField
-                          $valid={errors.website && touched.website}
-                          name="website"
-                          placeholder={data.form.website}
-                        />
-                        <F.ErrorWrapper>
-                          <F.FormikError component="div" name="website" />
-                        </F.ErrorWrapper>
-                      </F.Group>
-                      <F.Group>
-                        <SelectWrapper>
-                          <F.Label htmlFor="amr">
-                            {data.form.select.title}
-                          </F.Label>
-                          <F.Select
-                            name="amr"
-                            id="amr"
-                            defaultValue={"DEFAULT"}
-                            onChange={handleChange}
-                            onBlur={handleBlur}
-                            required
-                          >
-                            <option value="DEFAULT" disabled>
-                              {data.form.select.default}
-                            </option>
-                            {data.form.select.options.map((item) => (
-                              <option key={item} value={item}>
-                                {item}
-                              </option>
-                            ))}
-                          </F.Select>
-                        </SelectWrapper>
-                      </F.Group>
-                      <F.Group>
-                        <Button
-                          disabled={isSubmitting || !isValid || !dirty}
-                          type="submit"
-                          variant="primary"
-                          size="large"
-                        >
-                          {data.form.btn}
-                        </Button>
-                      </F.Group>
-                    </F.FormikForm>
-                  )}
-                </Formik>
-              </FormWrapper>
+              <GetFundedForm data={data} language="us" />
             </RightContentWrapper>
           </RightWrapper>
         </ContentWrapper>
@@ -344,10 +174,6 @@ const PageTitle = styled(T.H1)`
 const Title = styled(T.H2)`
   margin-bottom: 8px;
 `
-const SelectWrapper = styled(F.Group)`
-  margin-top: 16px;
-  /* margin-bottom: 40px; */
-`
 const Description = styled(T.Body)``
 
 const TextWrapper = styled.div`
@@ -392,13 +218,5 @@ const ContentWrapper = styled.div`
 
   @media (max-width: 800px) {
     grid-template-columns: 1fr;
-  }
-`
-const FormWrapper = styled.div`
-  margin-top: 24px;
-
-  .ButtonWrap button {
-    width: 100%;
-    margin: 16px 0 0 0;
   }
 `
